whatsapp-bridge/requirements.txt

fastapi==0.111.0
uvicorn==0.30.1
selenium==4.21.0
webdriver-manager==4.0.1
python-dotenv==1.0.1
requests==2.32.3
